+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.0-beta.8
 - core: Update dio to 4.0.0-prev1.
 - fix: update cache header values on 304 (keeping in sync freshness of our cache content).
+- fix: return cache on socket exception with hitCacheOnErrorExcept.
 
 ## 2.0.0-beta.7
 - core: Update dio to 4.0.0-beta7.